+++ routes/userbackend.php
@@ -5,6 +5,7 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\User\HomeController;
 use App\Http\Controllers\User\TicketController;
+use App\Http\Controllers\User\KnowledgebaseController;
 
 Route::get('/backend/user/getProfile', [ProfileController::class, 'backend_getUserProfile']);
 
@@ -22,3 +23,8 @@ Route::get('/backend/user/getAuditTrails/{ticket_number}', [TicketController::cl
 Route::get('/backend/user/checkTicketRatings/{ticket_number}', [TicketController::class, 'backend_checkTicketRatings']);
 Route::post('/backend/user/addTicketComment', [TicketController::class, 'backend_addTicketComment']);
 Route::post('/backend/user/addTicketRating/{ticket_number}', [TicketController::class, 'backend_addTicketRating']);
+
+// Knowledgebase
+Route::get('/backend/user/getFolders', [KnowledgebaseController::class, 'backend_getFolders']);
+Route::get('/backend/user/getTopics/{folder_id}', [KnowledgebaseController::class, 'backend_getTopics']);
+Route::get('/backend/user/getTopicInfo/{topic_id}', [KnowledgebaseController::class, 'backend_getTopicInfo']);